fix: tags controller n-n relationships to posts

+++ app/Http/Controllers/CategoriesController.php
@@ -98,6 +98,10 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if ($category->posts()->count() > 0) {
+            session()->flash('error', 'Posts found : ' .$category->posts()->count(). '. Category cannot be deleted because it has posts.'); 
+            return redirect()->back();
+        }
         $category->delete();
 
         session()->flash('success', 'Category deleted successfully');
