+++ app/Http/Controllers/TagsController.php
@@ -98,6 +98,10 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if ($tag->posts()->count() > 0) {
+            session()->flash('error', 'Posts found : ' .$tag->posts()->count(). '. Tag cannot be deleted because association to some posts exist.'); 
+            return redirect()->back();
+        }
         $tag->delete();
 
         session()->flash('success', 'Tag deleted successfully');
